+++ project/package.json
@@ -5,6 +5,8 @@
   "dependencies": {
     "@reduxjs/toolkit": "^1.8.6",
     "axios": "^0.27.2",
+    "classnames": "^2.3.2",
+    "dayjs": "^1.11.6",
     "history": "^5.3.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
