+++ project/src/components/footer/footer.tsx
@@ -1,9 +1,9 @@
+import Logo from '../logo/logo';
+
 function Footer(): JSX.Element {
   return (
     <footer className="footer container">
-      <a className="footer__logo-link" href="main.html">
-        <img className="footer__logo" src="img/logo.svg" alt="6 cities logo" width="64" height="33" />
-      </a>
+      <Logo type='footer' />
     </footer>
   );
 }
