+++ project/src/components/map/map.tsx
@@ -1,0 +1,13 @@
+import cn from 'classnames';
+
+type MapProps = {
+  className: string;
+}
+
+function Map({className}: MapProps): JSX.Element{
+  return (
+    <section className={cn(className, 'map')} />
+  );
+}
+
+export default Map;
