+++ project/src/components/tabs/tabs.tsx
@@ -1,0 +1,42 @@
+function Tabs(): JSX.Element{
+  return (
+    <div className="tabs">
+      <section className="locations container">
+        <ul className="locations__list tabs__list">
+          <li className="locations__item">
+            <a className="locations__item-link tabs__item" href="/#">
+              <span>Paris</span>
+            </a>
+          </li>
+          <li className="locations__item">
+            <a className="locations__item-link tabs__item" href="/#">
+              <span>Cologne</span>
+            </a>
+          </li>
+          <li className="locations__item">
+            <a className="locations__item-link tabs__item" href="/#">
+              <span>Brussels</span>
+            </a>
+          </li>
+          <li className="locations__item">
+            <a className="locations__item-link tabs__item tabs__item--active" href="/#">
+              <span>Amsterdam</span>
+            </a>
+          </li>
+          <li className="locations__item">
+            <a className="locations__item-link tabs__item" href="/#">
+              <span>Hamburg</span>
+            </a>
+          </li>
+          <li className="locations__item">
+            <a className="locations__item-link tabs__item" href="/#">
+              <span>Dusseldorf</span>
+            </a>
+          </li>
+        </ul>
+      </section>
+    </div>
+  );
+}
+
+export default Tabs;
