+++ project/src/constants.ts
@@ -12,3 +12,48 @@ export enum AuthorizationStatus {
 }
 
 export const MAX_RATING = 5;
+
+export const LogoSize = {
+  header: {
+    width: 81,
+    height: 41
+  },
+  footer: {
+    width: 64,
+    height: 33
+  }
+};
+
+export const cardImageSize = {
+  favorites: {
+    width: 150,
+    height: 110
+  },
+  cities: {
+    width: 260,
+    height: 200
+  },
+};
+
+export const RATING = [
+  {
+    value: 1,
+    title: 'terribly'
+  },
+  {
+    value: 2,
+    title: 'badly',
+  },
+  {
+    value: 3,
+    title: 'not bad',
+  },
+  {
+    value: 4,
+    title: 'good',
+  },
+  {
+    value: 5,
+    title: 'perfect',
+  }
+];
