--- project/src/pages/not-found-page/not-found-page.tsx
@@ -1,20 +1,0 @@
-import styles from './not-found-page.module.css';
-import Header from '../../components/header/header';
-import Footer from '../../components/footer/footer';
-
-function NotFoundPage(): JSX.Element {
-  return (
-    <div className="page">
-      <Header />
-
-      <section className={styles.section}>
-        <h1 className={styles.title}>404. Page not found</h1>
-        <a href="/">Вернуться на главную</a>
-      </section>
-
-      <Footer />
-    </div>
-  );
-}
-
-export default NotFoundPage;
